requirements.txt
  flask
  ultralytics
  opencv-python-headless
  numpy
  pillow
  python-dotenv
  azure-ai-openai 